fix stories
